Generalize user-agent setup
