+++ sensor/upgrader/main.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"flag"
 
+	"github.com/stackrox/rox/pkg/clientconn"
 	"github.com/stackrox/rox/pkg/logging"
 	"github.com/stackrox/rox/pkg/utils"
 	"github.com/stackrox/rox/pkg/version"
@@ -33,6 +34,8 @@ func mainCmd() error {
 	if err != nil {
 		return err
 	}
+
+	clientconn.SetUserAgent("upgrader")
 
 	upgradeCtx, err := upgradectx.Create(context.Background(), upgraderCfg)
 	if err != nil {
